Bless tests.
