+++ src/test/ui/issues/issue-3214.rs
@@ -5,8 +5,7 @@ fn foo<T>() {
 
     impl<T> Drop for Foo<T> {
         //~^ ERROR this struct takes 0 generic arguments but 1 generic argument
-        //~| ERROR the type parameter `T` is not constrained by the impl trait, self type, or predicates
         fn drop(&mut self) {}
     }
 }
-fn main() { }
+fn main() {}
