+++ src/test/ui/stability-attribute/generics-default-stability.rs
@@ -13,18 +13,6 @@ impl Trait1 for S {
 
 struct S;
 
-impl Trait1<usize> for S { //~ ERROR use of unstable library feature 'unstable_default'
-    fn foo() -> usize { 0 }
-}
-
-impl Trait1<isize> for S { //~ ERROR use of unstable library feature 'unstable_default'
-    fn foo() -> isize { 0 }
-}
-
-impl Trait2<usize> for S { //~ ERROR use of unstable library feature 'unstable_default'
-    fn foo() -> usize { 0 }
-}
-
 impl Trait3<usize> for S {
     fn foo() -> usize { 0 } // ok
 }
